Upgrade typescript-eslint and fix new errors/warnings
This upgrade introduced around 20 errors and 30 warnings.

The errors were mostly related to using "banned" types such as `object |
Function`. I've fixed all the errors.

The warnings were mostly

> Missing return type on function. (@typescript-eslint/explicit-module-boundary-types)

I've fixed all but one: the return type of useToolTip (which looks insane). I'm not sure what to do with this one, and want to discuss with the team.

We may want to suppress the `explicit-module-boundary-types` warning anyway?

Overall, I'd say this upgrade improves the code quality in general.